move all nodes to a new module/directory
